Suggestion for pThumb/Babel issue
See: http://forums.modx.com/thread/88002/pthumb-and-multilanguage-site

At the moment an empty 'phpthumbof.cache_url' but existing system setting is not used (because skipEmpty is true in getOption). In Babel the 'base_url' setting is changed context specific, but the 'phpthumbof.cache_url' has to point to the MODX base url (empty 'phpthumbof.cache_url' setting). 

By moving the closing bracket of the rtrim and filling the setting with a space character the 'phpthumbof.cache_url' setting is used and the rtrim removes the space (before it don't because of the concatenated '/'). Changing the code this way should cause no side effect as changing skipEmpty to false in getOption would do.